Make project card divs have left-side and ..
right-side classes so that textdivs and iconsdivs can be
selected as children of the appropriate side. Reduce padding On
right side of project-cards to increase padding-right on icons, keeping
them in the same place, to enable text to go closer to right edge of
cards on the right-side of the page.